Replace from_global_id by get_node

+++ users/schemas/mutations/research.py
@@ -1,10 +1,10 @@
 from graphene import relay, Field, List, String, Int, ID
-from graphql_relay.node.node import from_global_id
 
 from danesh_boom.viewer_fields import ViewerFields
 from users.schemas.queries.research import ResearchNode
 from users.models import Research
 from users.forms import ResearchForm
+from utils.relay_helpers import get_node
 
 
 class CreateResearchMutation(ViewerFields, relay.ClientIDMutation):
@@ -51,11 +51,14 @@ class UpdateResearchMutation(ViewerFields, relay.ClientIDMutation):
     @classmethod
     def mutate_and_get_payload(cls, input, context, info):
         user = context.user
-        id = input.get('id', None)
-        research_id = from_global_id(id)[1]
-        research = Research.objects.get(pk=research_id)
+        research_id = input.get('id', None)
+        research = get_node(research_id, context, info, Research)
+
+        if not research:
+            raise Exception("Invalid Research")
+
         if research.user != user:
-            raise Exception("Invalid Access to Work Research")
+            raise Exception("Invalid Access to Research")
 
         # update research
         form = ResearchForm(input, instance=research)
@@ -77,11 +80,14 @@ class DeleteResearchMutation(ViewerFields, relay.ClientIDMutation):
     @classmethod
     def mutate_and_get_payload(cls, input, context, info):
         user = context.user
-        id = input.get('id', None)
-        research_id = from_global_id(id)[1]
-        research = Research.objects.get(pk=research_id)
+        research_id = input.get('id', None)
+        research = get_node(research_id, context, info, Research)
+
+        if not research:
+            raise Exception("Invalid Research")
+
         if research.user != user:
-            raise Exception("Invalid Access to Work Research")
+            raise Exception("Invalid Access to Research")
 
         # delete research
         research.delete()
